@@ -3,14 +3,14 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@emotion/react": "^11.6.0",
-    "@emotion/styled": "^11.6.0",
+    "@emotion/react": "^11.11.4",
+    "@emotion/styled": "^11.11.5",
     "@mui/icons-material": "^5.2.0",
-    "@mui/material": "^5.2.1",
-    "@mui/system": "^5.2.1",
+    "@mui/material": "^5.15.20",
+    "@mui/system": "^5.15.20",
     "@reduxjs/toolkit": "^1.6.2",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "react-redux": "^8.0.5",
     "react-scripts": "^5.0.1"
   },
@@ -42,10 +42,10 @@
     "@testing-library/react": "^9.5.0",
     "@testing-library/user-event": "^7.2.1",
     "@types/jest": "^24.9.1",
-    "@types/node": "^12.20.55",
-    "@types/react": "^18.0.26",
-    "@types/react-dom": "^16.9.17",
-    "@types/react-redux": "^7.1.20",
+    "@types/node": "^18.19.39",
+    "@types/react": "^18.3.3",
+    "@types/react-dom": "^18.3.0",
+    "@types/react-redux": "^7.1.33",
     "@types/wicg-mediasession": "1.1.3",
     "babel-plugin-direct-import": "^1.0.0",
     "customize-cra": "^1.0.0",
