@@ -1,27 +1,28 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import { useState, useEffect, RefObject } from 'react'
 
-type IAudioContext = [
-  dataArray: Uint8Array,
-  audioBufferLength: number,
-  analyser?: AnalyserNode
-]
+type IAudioContext = [audioBufferLength: number, analyser?: AnalyserNode]
 
 // Initial analyser set up. Needs to only run once.
 export function useAudioContext(
   audioEl: RefObject<HTMLAudioElement>
 ): IAudioContext {
-  const [dataArray, setDataArray] = useState<Uint8Array>(new Uint8Array())
   const [analyser, setAnalyser] = useState<AnalyserNode | undefined>(undefined)
   const [audioBufferLength, setAudioBufferLength] = useState<number>(0)
+
+  let source: MediaElementAudioSourceNode
+  let audioContext: AudioContext
 
   useEffect(() => {
     if (!audioEl.current) return
 
-    const audioContext = new (AudioContext ||
-      (window as any).webkitAudioContext)()
+    if (!audioContext) {
+      audioContext = new (AudioContext || (window as any).webkitAudioContext)()
+    }
 
-    const source = audioContext.createMediaElementSource(audioEl.current)
+    if (!source) {
+      source = audioContext.createMediaElementSource(audioEl.current)
+    }
 
     // To ensure track actually plays
     audioEl.current.onplay = () => audioContext.resume()
@@ -34,12 +35,9 @@ export function useAudioContext(
     audioAnalyser.connect(audioContext.destination)
 
     const bufferLength = audioAnalyser.frequencyBinCount
-    // Empty array will be filled with frequency data
-    const data = new Uint8Array(bufferLength)
 
     setAudioBufferLength(bufferLength)
     setAnalyser(audioAnalyser)
-    setDataArray(data)
 
     return () => {
       // Cleanup and disconnect
@@ -48,5 +46,5 @@ export function useAudioContext(
     }
   }, [])
 
-  return [dataArray, audioBufferLength, analyser]
+  return [audioBufferLength, analyser]
 }
